+++ packages/firestore/src/api/database.ts
@@ -55,6 +55,7 @@ import {
   enableNetwork,
   ensureFirestoreConfigured,
   FirebaseFirestore as ExpFirebaseFirestore,
+  useFirestoreEmulator,
   waitForPendingWrites
 } from '../exp/database';
 import { FieldPath as ExpFieldPath } from '../exp/field_path';
@@ -95,7 +96,6 @@ import {
   updateDoc,
   Unsubscribe
 } from '../exp/reference_impl';
-import { DEFAULT_HOST } from '../exp/settings';
 import {
   DocumentChange as ExpDocumentChange,
   DocumentSnapshot as ExpDocumentSnapshot,
@@ -118,7 +118,7 @@ import {
   validateIsNotUsedTogether,
   validateSetOptions
 } from '../util/input_validation';
-import { logWarn, setLogLevel as setClientLogLevel } from '../util/log';
+import { setLogLevel as setClientLogLevel } from '../util/log';
 
 import { Blob } from './blob';
 import { LoadBundleTask } from './bundle';
@@ -236,17 +236,7 @@ export class Firestore
   }
 
   useEmulator(host: string, port: number): void {
-    if (this._delegate._getSettings().host !== DEFAULT_HOST) {
-      logWarn(
-        'Host has been set in both settings() and useEmulator(), emulator host will be used'
-      );
-    }
-
-    this.settings({
-      host: `${host}:${port}`,
-      ssl: false,
-      merge: true
-    });
+    useFirestoreEmulator(this._delegate, host, port);
   }
 
   enableNetwork(): Promise<void> {