+++ packages/firestore/src/local/memory_bundle_cache.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { Bundle, NamedQuery } from '../core/bundle';
+import { BundleMetadata, NamedQuery } from '../core/bundle';
 import {
   NamedQuery as ProtoNamedQuery,
   BundleMetadata as ProtoBundleMetadata
@@ -31,7 +31,7 @@ import { PersistencePromise } from './persistence_promise';
 import { PersistenceTransaction } from './persistence_transaction';
 
 export class MemoryBundleCache implements BundleCache {
-  private bundles = new Map<string, Bundle>();
+  private bundles = new Map<string, BundleMetadata>();
   private namedQueries = new Map<string, NamedQuery>();
 
   constructor(private serializer: LocalSerializer) {}
@@ -39,7 +39,7 @@ export class MemoryBundleCache implements BundleCache {
   getBundleMetadata(
     transaction: PersistenceTransaction,
     bundleId: string
-  ): PersistencePromise<Bundle | undefined> {
+  ): PersistencePromise<BundleMetadata | undefined> {
     return PersistencePromise.resolve(this.bundles.get(bundleId));
   }